Use custom fetchJson fn that adds proper headers
