+++ src/api/browse.ts
@@ -1,28 +1,10 @@
 import { searchResponseModel, datasetEmbeddedModel, datasetSummaryModel } from "../models/dataset";
 import { facetFilterModel } from "../models/facets";
-import authService from "../services/auth";
+import { fetchJson } from "../utils/utils";
+
 
 const SEARCH_URL = process.env.REACT_APP_SVC_SEARCH_URL;
 const REPOSITORY_URL = process.env.REACT_APP_SVC_REPOSITORY_URL;
-const CLIENT_URL = process.env.REACT_APP_CLIENT_URL;
-
-const getHeaders = async () => {
-  const user = await authService.getUser();
-  const headers: HeadersInit = {
-    "Accept": "application/json",
-    "Content-Type": "application/json"
-  };
-  if (CLIENT_URL) {
-    headers["Origin"] = CLIENT_URL;
-  }
-  const authorization = user?.access_token;
-  if (authorization) {
-    // the Authorization header is already used for Basic auth,
-    // therefore we use the X-Authorization header for the OIDC token
-    headers["X-Authorization"] = authorization;
-  }
-  return headers;
-}
 
 type getDatasetsSearchRespType = (
   callbackFunc: (hits: searchResponseModel) => void,
@@ -43,10 +25,9 @@ export const querySearchService: getDatasetsSearchRespType = async (
 ) => {
   let url = `${SEARCH_URL}/rpc/search`;
   url += `?document_type=${documentType}&return_facets=true&skip=${skip}&limit=${limit}`;
-  const headers = await getHeaders();
-  const body = JSON.stringify({ query: searchKeyword, filters: filterQuery });
+  const payload = { query: searchKeyword, filters: filterQuery };
   try {
-    const response = await fetch(url, {method: "POST", headers, body});
+    const response = await fetchJson(url, "POST", payload);
     const data = await response.json();
     callbackFunc(data);
   } catch {
@@ -73,9 +54,8 @@ export const getDatasetDetails: getDatasetDetailsType = async (
 ) => {
   let url = `${REPOSITORY_URL}/datasets/${datasetId}`;
   url += `?embedded=${embedded}`;
-  const headers = await getHeaders();
   try {
-    const response = await fetch(url, {method: "get", headers});
+    const response = await fetchJson(url);
     const data = await response.json();
     callbackFunc(data);
   } catch {
@@ -93,9 +73,8 @@ export const getDatasetSummary: getDatasetSummaryType = async (
   callbackFunc
 ) => {
   const url = `${REPOSITORY_URL}/dataset_summary/${datasetId}`;
-  const headers = await getHeaders();
   try {
-    const response = await fetch(url, {method: "get", headers});
+    const response = await fetchJson(url);
     const data = await response.json();
     callbackFunc(data);
   } catch {
